import_clean_socal_data_to_bigquery/requirements.txt
# Function dependencies, for example:
# package>=version
google-cloud
google-cloud-bigquery